+++ Binary-Strings/.gitignore
@@ -1,2 +1,3 @@
-/bin/
-/target/
+target/
+.settings/
+bin/
