+++ Binary-Strings/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>Binary-Strings</name>
+	<name>Tools-Utilities</name>
 	<comment></comment>
 	<projects>
 	</projects>
