+++ Binary-Strings/src/main/java/com/pwhiting/util/ClocData.java
@@ -39,7 +39,7 @@ public class ClocData {
 		return languageStats.values().toArray(new LangStats[languageStats.values().size()]);
 	}
 
-	Map<Language, LangStats> getLanguageStatsMutable() {
+	public Map<Language, LangStats> getLanguageStatsMutable() {
 		return languageStats;
 	}
 
@@ -49,7 +49,7 @@ public class ClocData {
 	 * @param data
 	 * @return
 	 */
-	ClocData imprint(final ClocData data) {
+	public ClocData imprint(final ClocData data) {
 		header.imprint(data.header);
 		languageStats.clear();
 		languageStats.putAll(data.languageStats);
@@ -67,9 +67,20 @@ public class ClocData {
 
 	}
 
-	void reset() {
+	public void reset() {
 		header.reset();
 		languageStats.clear();
+	}
+
+	/**
+	 * Get number of files that use the specified language.
+	 *
+	 * @param lang
+	 * @return count
+	 */
+	public LangStats getLangStats(final Language lang) {
+		return this.getLanguageStatsMutable().containsKey(lang) ? this.getLanguageStatsMutable()
+				.get(lang).copy() : new LangStats(lang);
 	}
 
 	@Override
@@ -147,7 +158,7 @@ public class ClocData {
 		 * @param header
 		 * @return
 		 */
-		Header imprint(final Header header) {
+		public Header imprint(final Header header) {
 			clocUrl = header.clocUrl;
 			clocVersion = header.clocVersion;
 			elapsedSeconds = header.elapsedSeconds;
