+++ Binary-Strings/src/main/java/com/pwhiting/util/Util.java
@@ -2,7 +2,9 @@ package com.pwhiting.util;
 
 import java.lang.reflect.Field;
 import java.lang.reflect.Modifier;
+import java.util.Arrays;
 import java.util.Date;
+import java.util.List;
 import java.util.Map;
 
 import org.slf4j.Logger;
@@ -10,6 +12,7 @@ import org.slf4j.LoggerFactory;
 
 import ch.qos.logback.classic.Level;
 
+import com.google.common.collect.Lists;
 import com.google.common.collect.Range;
 
 /**
@@ -238,6 +241,59 @@ public final class Util {
 		value.append("]");
 
 		return value.toString();
+
+	}
+
+	public static Byte[] arrayConversion(byte[] array) {
+
+		Byte[] value = new Byte[array.length];
+
+		for (int i = 0; i < value.length; i++) {
+			value[i] = array[i];
+		}
+
+		return value;
+	}
+
+	public static byte[] arrayConversion(Byte[] array) {
+
+		byte[] value = new byte[array.length];
+
+		for (int i = 0; i < value.length; i++) {
+			value[i] = array[i];
+		}
+
+		return value;
+	}
+
+	public static List<Byte[]> byteArraySplit(byte[] value, byte[] pattern) {
+
+		List<Byte[]> split = Lists.newArrayList();
+
+		if (value.length >= pattern.length) {
+
+			int charLength = pattern.length;
+
+			int lastSplit = 0;
+
+			for (int i = 0; i < value.length - pattern.length; i++) {
+
+				byte[] selection = Arrays.copyOfRange(value, i, i + charLength);
+
+				if (Arrays.equals(pattern, selection)) {
+					i += charLength - 1;
+					split.add(arrayConversion(Arrays.copyOfRange(value,
+							lastSplit, i - charLength + 1)));
+					lastSplit = i + 1;
+				}
+
+			}
+
+			split.add(arrayConversion(Arrays.copyOfRange(value, lastSplit,
+					value.length)));
+		}
+
+		return split;
 
 	}
